% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otsu.R
\name{make_bw}
\alias{make_bw}
\title{Changes a grayscale image to black and white}
\usage{
make_bw(image)
}
\arguments{
\item{image}{The image generated after the ExG transformation in the EBImage format.}
}
\value{
A binary image
}
\description{
Takes in a grayscale image and finds the best threshold for binarization of the image.
}
