% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotations}
\alias{rotations}
\title{Rotates an image by x degrees}
\usage{
rotations(picture, degrees)
}
\arguments{
\item{picture}{The image to be rotated.}

\item{degrees}{The degrees of rotation until 180 (30 = every 30 degrees).}
}
\value{
A list of pictures that have been rotated by x degrees each.
}
\description{
Takes an image and rotates it by a number of degrees choosen by the user.
}
\details{
The function takes in an image and rotates it by a number of degrees chosen by the user. It will keep rotating until it has reached the 360 degree limit and save the images into a list.
}
