% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smooth the average of binary picture column values}
\usage{
smoothing(picture, intensity)
}
\arguments{
\item{picture}{The binary image.}

\item{intensity}{The intensity of smoothening of the vector.}
}
\value{
A smoothed vector.
}
\description{
Reads in an image and takes the average column value then smooths the array for more defined local maximas and minimas.
}
