% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{blobify}
\alias{blobify}
\title{Denoise a black and white image to only core features}
\usage{
blobify(images, size)
}
\arguments{
\item{images}{One, or more, image(s).}

\item{size}{A number, the starting size of the kernel for morphology (minimum of 2).}
}
\description{
This takes a black and white image, preferrably the image produced by an otsu transformation, and gets rid of all extrenuous features. Many long skinny features and random clumps of features will be erased, leaving only the larger core features of an image.
}
